requirements.txt
pyinstaller
PyGObject
python-dotenv